Add better error handling
A custom Error type is created and the main body of execution is wrapped
in a function that returns a Result instead of panicking on an error.

Better error messages are still needed. Ideally, the output from stderr
should be used for the error message for the build, compile, link, and
sign steps.