Fix GUIDs in template
The `UpgradeCode` and `Guid` for the PATH environment variable needed to
be replaced with placeholders. Eventually the mustache-rust and guid
crates should be used to automatically replace the placeholders with
GUIDs each time the `--print-template` flag is used so users can begin
using contents immediately without manual modification. This will
dovetail nicely into the eventual `--init` flag.