Remove `--x64` flag
The `--x64` flag is replaced with a conditional compilation (`cfg!`)
that checks the target architecture on which the subcommand was
installed. Thus, if the `x86_64` toolchain is used, then the platform
will default to `x64` in the installer, but if the `i686` toolchain is
used, then the platform will default to `x86` in the installer.

An override flag/option might be needed in the future, but this makes
things more automatic. It still needs to be tested.