Add automatic GUID generation on printing of template
When the `--print-template` flag is used, GUIDs are automatically
generated for the `UpgradeCode` attribute of the `Package` element and
the `Guid` attribute of the Path component. Each time the
`--print-template` flag is used, a new set of GUIDs are generated. This
makes it a lot easier to redirect the output to a file and not have to
modify the WiX source (wxs) file.