Add `-l,--license` option
A `License.txt` file is added to the install location along side the
`bin` folder by the installer. The `License.txt` file is created from
the path specified in the `license-file` field in the package's
manifest or from a `LICENSE` file in the root project folder if the
`license-file` is not used automatically. The `-l,--license` option
overrides this default behavior. Note, the file specified with the
`-l,--license` option is still converted to `License.txt` during the
creation of the installer regardless of the format of the file.

This is an initial implementation for handling licenses for
installation, but the license that appears _during_ installation still
needs to be implemented.