Add timestamp server aliases
Aliases are added for `Comodo` and `Verisign`. If these are specified as
the value for the `-t,--timestamp` option, then the hard-coded URLs for
either of these aliases are used. This way the user does not need to
remember the URLs.

A brief search for all timestamp servers only yielded these two in
existence. There might be more, but the framework for adding timestamp
server aliases has been implemented within the code.