Add `-B,--bin-path` option
It is not fully implemented, but it has been added to the interface and
includes a test. I have decided to rename the `-W,--wix-path` option to
`-B,--bin-path` option because this specifies the path for the _bin_
folder, not the wix folder. It also looks a little weird with the full
command, i.e. `cargo wix --wix-path` is just too many wixes.