Add license templates and file generation
The `--init` flag now generates a `wix\License.rtf` file based on the
`license` field in the package's manifest (Cargo.toml). The manufacturer
is used for the copyright holder and the package name is used for the
product name unless overriden by the environment or command line. The
copyright year still needs work as I am not sure where to pull this from
by default (current year?).