Remove `--print-wxs` and `--print-license`
Instead of having different arguments to print the various templates,
a single `--print-template` option is used. The printing of the WXS and
licenses are combined. This makes the CLI a lot cleaner. There is no
real need for separate options/flags for printing the templates. All of
them are templates and the user does not really care which one is used
for which purpose when trying to print.