Remove License type
There was too much overlap between the License and Template types. So,
the License type was removed and there is now just the Template type.
Depending on the variant, different implementation are executed for
a template. This makes the Template type easier to add to in the future
as well.