Change `run` to `build`
The `--print-template` option, `--init` flag, etc. also "run", so having
the `run` method for the builder API of the Wix struct is a little
ambiguous. Thus, it is changed to `build` because it is building the
project and installer. It could also be called `create`, but let's start
with `build` and see how it works.