Change `build` method to `create` method
The `build` method is slightly confusing because the first step is to
build the project use the release profile, then build/create the
installer, and possibly sign it. Eventually, a `--no-build` flag will be
added, which will skip the project build step (`cargo build --release`).
The `--no-build` flag would seem to mean to skip the `build` method of
the Wix struct, when it is just skipping the project build, not the
installer build. So, the `build` method is changed to the `create`
method because it is ultimately creating an installer and reserves the
"build" as a step in the creation process.