Revert to enabling ANSI escape sequences
The `loggerv` crate uses the `ansi_term` crate for its coloring. This
means that we still need to enable ANSI escape sequences to prevent
printing ANSI escape characters in the logging output. It might be worth
changing from the loggerv crate to the env_logger crate. The env_logger
crate also uses the termcolor crate for colorized output instead of the
`ansi_term` crate and it would provide support for setting the log level
from an environment variable.