Change to env_logger
The `loggerv` crate is replaced with the `env_logger` crate and
a modified format with level, or filter, set via verbosity. The
functionality of the verbosity flag, format, and coloring should be
close to the implementation with the `loggerv` crate except for the
following:

1. Line numbers are never printed. This did not appear to be very useful
without the module paths. The logging is being used for end-user
information sharing, not for debugging or developer usage, so the line
numbers and module paths are useless to end-users.
2. TRACE and DEBUG statements are both white but differ in boldness.
Light Gray and Gray are not supported on non-ANSI capable Windows
consoles. The color selection is limited.
3. Using a verbosity of greater than three has no effect on formatting
or filters.
4. Environment variables are not currently supported as there remains
many questions about expected functionality and implementation.