Add checks for existence of compiler
The path to the `bin` folder for the compiler of the WiX Toolset is
checked for existence before starting the compiler. A better error
message is provided if the path does not exist and/or the compiler
cannot be found than the standard, generic IO error message.

This should possibly resolve:

- [Issue #46](https://github.com/volks73/cargo-wix/issues/46)