Add better error messages if PATH fails
If the compiler and linker applications could not be found in the PATH
environment variable, which is the default behavior, the sub-command
would emit a generic file not found OS Error (2), which is an IO error.
This is too generic and may be the source of Issue #46.

A better error message is provided, but the only way to know if this
worked is to attempt the compiler and/or linker command and then map the
IO error if the kind is `NotFound` to a Generic error.