Change WIX_PATH to the WIX environment variable
The `WIX` environment variable is created for the system during the
installation of the WiX Toolset v3.11. This points to the root folder of
the installation, typically `C:\Program Files (x86)\WiX Toolset v3.11`.
This serves a very similar purpose to the `WIX_PATH` environment
variable. It is more robust because it will not necessarily exist if the
the WiX Toolset is not installed, which makes error resolution easier.
See [Issue #53](https://github.com/volks73/cargo-wix/issues/53).