Change some options to require the `--init` flag
The `-d,--description`, `-m,--manufacturer`, and `-p,--product-name`
options are changed to require the `--init` flag. These options are only
needed when generating a new WiX Source file to populate the embedded
template. Eventually the XML preprocessor variables will be removed and
template variables will be used instead. After the wxs file has been
generated as part of the initialization, the values of these options can
be changed in the wxs file with a text editor. The thought being that
these values are not changing each time an installer is created, but
only rarely when a project changes hands or direction.