Change sidecar license file to be optional
If the `license-file` field is _not_ set in the package's Cargo.toml
manifest and the `--license` option is also not used during
initialization, then the sidecar license is not added to the WXS file
and therefore not included in the installer. This eliminates compiler
errors for the candle.exe application when license files could not be
found.