Add `initialize` module
The `initialize` module will embody all of the code and implementation
details for the initialization "mode" of the subcommand. This includes
its own Builder and Executor types. This helps eliminate the "get_*"
methods and better organizes the code into the Builder pattern. Many
other features are being added as this refactoring continues.

A separate module with a Builder and Executor will eventually be added
for each mode: create and print.