Add Eula type
The EULA type indicates the origin and/or kind of EULA. Does a license
file with the RTF extension already exist to be used as the EULA? Was
the license file specified on command line or in the manifest? Does
a EULA need to be generated? The EULA type answers these questions.