Change struct name
The `Initialize` struct is renamed to `Execution`. Each subcommand/mode
will get a Builder and Execution type that will be in a sub-module of
the mode. For example, a `creation` module will be added that will have
a `Builder` and `Execution` type that creates the installer.
A `signature` module will be added for signing installers that will also
have the Builder/Execution types. Since the module names will indicate
modes, the `Initialize` type/name seemed redundant.