Add integration test for `license` field
This does not fully pass yet because the default appears to be to not include a
sidecar, even if the license field value is recognized and the RTF file is
generated. I believe the sidecar file should be included if the license is
generated. This might be an order of operations thing that I need to look into.