Fix default behavior for license during initialization
The implementation for determining if a sidecar license file should be used when
the `license` field in the package's manifest (Cargo.toml) contains a supported
license is fixed so the integration tests pass. If the license ID is supported,
then the sidecar license file will be included in the installer.