Add integration tests
Some of these tests currently fail. These integration test actually build, or
create, installers using the WiX Toolset. The integration test fail because the
WXS file that is generated causes errors in either the candle.exe (compiler) or
light.exe (linker) applications. In most cases, the failures are related to
paths and file existence checks. These probably need to be eliminated from the
initialize mode. In other cases, it is implementation problem that needs to be
resolved. The tests are committed and pushed as a marker of where work is needed.