Add failing integration test
A creation test is added that currently fails. This tests if the license,
documentation, description, homepage, and repository fields for the temporary
package's manifest exist does the installer build without error using no
initialization options.

It currently fails because the path to the generated license file (License.rtf)
cannot be found. I am guessing this is due to a relative path issue.