Fix `binary-name` not working
The `binary-name` related options and methods have been changed to just
`binary` and it now works similar to the `license` option and methods. The
binary name and binary source are derived from the binary path passed to the
`binary` option or built dynamically based on the package's
manifest (Cargo.toml). This is more consistent with rest of the API and usage
and results in the tests passing.