Add test when input and output are used for initialization
This also required changing the path existence checks. It is possible that the
files do not exist at the time of initializing, but they will exist when the
user eventually creates the installer. For example, banner images, custom
licenses, custom EULA, etc. So, the existence checks should not actually be
here. If the file cannot be found, the WiX Toolset's linker (light.exe) will
error with a file not found and the `--nocapture` option can be used to
determine the exact error and file that is missing.