Change short flags and options for create mode
To be consistent with the init and print modes, the short flags and options for
the create mode are changed to all lowercase unless there would be duplicates.
If there would be a duplicate, then the flag or option first in alphabetical
order gets the lowercase short value and the second flag or option in
alphabetical order gets the uppercase short value. There is currently no
triplicates.