Change `holder` to `owner` for init and print modes
The `holder` option is renamed to `owner` for the init and print
subcommands/modes. This avoids having a `-h` short option that conflicts with
the help short message flag but allows the "owner" to have a mneumonic short
options, `-O`. Note, the `-o` is already taken but the `output` option, which is
convention.