Fix back button going to license agreement dialog
When no EULA is used, the `Back` button in the Customize dialog would go back to
the License Agreement dialog instead of the Welcome dialog, where placeholder
text would appear. The issue appears to be caused by too low of an order for the
`Back` button event in the WXS template.

See the comments to the [Removing License Dialog] StackOverflow Q&A for more
information about implementation and fix.

[Removing License Dialog]: https://stackoverflow.com/questions/13045658/removing-license-dialog