Bump test dependencies
This required changing the implementation of creating test projects. The `.tmp`
prefix can be used in Windows but the cargo project name cannot. Thus, the
`--name <NAME>` option needed to be used for creating a test package because the
`TempDir` type from the assert_fs package does not support the tempfile TempDir
builder or changing the prefix in newer versions. See Issues [#46] and [#48]
from the `assert_fs` project for more information.

[#46]: https://github.com/assert-rs/assert_fs/issues/46
[#48]: https://github.com/assert-rs/assert_fs/issues/48