Add initial implementation for subcommand configuration
The subcommand, `cargo wix`, can be configured for the _create_ (default)
subcommand using the package's manifest (Cargo.toml) and a
`[package.metadata.wix]` section. Every CLI option for the _create_ (default)
subcommand can be defined in the metadata section, but if the CLI option is
present, it will override the metadata section.