Change `inputs` to `includes` and implement `input` feature
The `inputs` feature is changed to `includes` and the `-I,--include` option is
added to the CLI. The `<INPUTS>` argument is changed to `<INPUT>`, not the
singular versus plural, and now accepts only a `Cargo.toml` file instead of
multiple WXS files.

The implementation for specifying a package's manifest file has been added, but
it has not been tested. This is just the implementation.