Fix initializing logging error when running multiple tests
The `init_logging` function can only be called once to setup test logging. This
is lazily executed via the TARGET_WIX_DIR constant. I am not sure where else to
put this. The `init_logging` function might be modified so it can be called
within in each test intead of the "hack" for once per the create integration
test module.