Add `-d,--dbg-build` and `-D,--dbg-name` flags
The two flags are added to the default _create_ subcommand and toggle building
the package with the Debug profile using Cargo and append `-debug` to the file
stem (content before the dot and file extension) for the installer's file name.
These should generally be used in combination without each other so that
end-users know the installer is for a debugging variant of the application that
does not include any optimizations.

These are broken into two separate flags because there might be situations for
developers to not want to rename the MSI file but build with debugging
information, or more likely, the application was built with the Release profile
that contains optimizations but other mechanisms are included to allow for
debugging (logging, reporting home, etc) and the developer wants to provide two
variants.