Fix default binaries source paths with different profile
Debug builds are located in the `target\debug` folder instead of the
`target\release` folder. The `target\release` folder was hard-coded in may
places, but it was especially problematic because WXS file created with the
`cargo wix init` command would generate the binary source paths with the
`release`. Creating the installer is where the profile is selected, so a
preprocessor variable needed to be added to make this work, such that the
profile, "debug" or "release", could be "passed" to the WXS file for each
default binary source path.