Add `InstallerKind` and `WixObjKind` enum types
The `InstallerKind` is broken into two separate enum types: (i) `WixObjKind`
and (ii) `InstallerKind` because a single "Installer Kind" type was doing double
duty as a type for the identifying a WiX object (wixobj) file and then
indicating the type of installer to be created. The `WixObjKind` now indicates
the type of WiX object file and it is used to determine the kind of
installer (exe or msi) to create.

This division might appear to be more verbose, but I believe it will be easier
to follow and maintain in the future. First, identify the kind of WiX object for
all object files created from the compiler (candle), then from the kinds of WiX
object files, determine the kind of installer to create. This also makes writing
tests easier.