Add information about minimum version of Rust
The use of the `Option::as_deref()` method requires a minimum version of Rust to
be v1.40.0 or newer. At minimum then, v1.40.0 is needed to compile this
sub-command with the latest features and code. The CI configuration if updated
to use the minimum version of Rust as well as a method for indicating supported
versions and a version is added to the installation requirements in the README.

There is also the use of the `TryFrom` trait, which was introduced in v1.34.0.