Add automatic inclusion of the WixBalExtension for a bundle
Some empirical testing with creating a very simple bundle outside of a Rust
project indicates that the `WixBalExtension` needs to be included with the
linker invocation if the WiX standard bootstrapper application is used, which is
the most common use case. A reference to a bootstrapper application is required
for a bundle. This will automatically take care of including the extension when
a bundle-based installer is detected.