Remove unnecessary log statement
The trace statement does not really provide useful information. This
method is always executed and we are always going to use the package's
manifest file path to find all WXS files. I think this originally had
the path as part of the log statement and was useful when debugging and
testing that the correct file path was being used. With the new
cargo-metadata implementation, I think there is no utility to this log
statement.