Add upgrade code to printer
The WXS printer is used with the `cargo wix init` command, i.e. the
initialize module, to generate the WXS file. Thus, if the initialize
module and command are going to have an upgrade code option, then so
does the WXS printer and `cargo wix print` command. In other words,
implementing an upgrade code option for the init command starts with
implementing it in the print module and command.