Refactor implementation for consistency
The implementation for determining the target string appeared to be
split between the `target` method and the `get_host_triple` function.
The `target` method was relatively simple with the `get_host_triple`
function doing most of the work. The `get_host_triple` function was a
private function to the module and not used anywhere else outside of the
`target` method. The implementation from the `get_host_triple` is moved
to the `target` method and removed.

The handling of the pattern matching with `Option` type is changed to
match (pun) the implementation in other similar methods.