Add documentation about static linking CRT
A section and heading is added to the binary module-level documentation
with details and information about the CRT dependency when using the
MSVC runtime and static linking with the Rust compiler (rustc).

Efforts are underway to make this less of an issue, but for the time
being some recognition of the issue and information on resolution in the
documentation would probably be appreciated by users.