Fix always setting the build target
Always setting the build target, `--target=<TRIPLE>`, will always place
the build artifacts for the rustc/Cargo build step in a subfolder of the
`target` folder with the name of the triple, i.e.
`target\<TRIPLE>\release`. If no target triple is specified, then the
build artifacts are located in `target\release`. The target triple
should only be passed to the cargo builder if it was explicitly set;
otherwise, the default "system" target should be used and the
`target\release` folder will work.

This was causing all but the cross-compilation test to fail.