Change to 2018 edition imports
The `extern` syntax is changed to the 2018 edition for the binary and
library code. This mostly involved removing the extern lines, with one
exception, and using the `use log::{debug, info, trace, warn}` syntax
for the log macros.

I am investigating how to use CFG attributes for imports to eliminate
the last remaining extern one.