Add using CARGO environment variable
If building the project, then the path defined in the `CARGO`
environment variable should be used. If the `CARGO` environment variable
is not defined or empty, then default to the `cargo` name with
executable defined in the `PATH` environment variable for the system.
This is similar to the implementation found in other crates and cargo
subcommands. See Issue #136.