add --target-bin-dir to execute
This sets CargoTargetBinDir in the wix template, allowing something orchestrating
cargo-wix to take the reigns on building/gathering binaries completely. Setting
this an --no-build supresses the warning (as at this point there's a pretty
clear signal the user is doing this intentionally and not as a temp thing).

No integration tests were added for this functionality because it wasn't clear
to me how to reasonable test this, but the functionality is boring enough that
it's easy to trust.

As compensation some backlog tests/docs were added.

-----

Also does some cleanups to factor target-triple/profile logic so that things
aren't recomputed in a bunch of places.

Also adds the arguably-forgotten-to-be-added package.metadata.wix.profile
(for projects that always want to build with a special profile).

Note that --target-bin-dir this time is *intentionally* without a package.metadata
pair, because it's pretty exceptional to want to fully bake the target dir.